Recipie Bloc Logic Finished

+++ lib/bloc/recipie/recipie_bloc.dart
@@ -17,10 +17,10 @@ class RecipieBloc extends Bloc<RecipieEvent, RecipieState> {
   RecipieBloc(
       {this.mealBloc, this.recipieRepository, this.recipieItemRepository})
       : super(RecipieLoading()) {
-    final recipieId = (state as RecipieLoadSuccess).recipieId;
+    final recipieId = (state as RecipieLoadSuccess).recipie;
     mealTrackGroupSubscription = mealBloc.listen((state) {
       if (state is MealLoadSuccess) {
-        add(LoadRecipieMeals(recipieId));
+        add(LoadRecipieMeals(recipieId.id));
       }
     });
   }
@@ -32,22 +32,24 @@ class RecipieBloc extends Bloc<RecipieEvent, RecipieState> {
     if (event is LoadRecipieMeals) {
       yield* _loadRecipie(event);
     } else if (event is AddMeal) {
-      yield* _addMeal(event);
-    } else if (event is UpdateMeal) {
-      yield* _editMeal(event);
+      yield* _addEditMeal(event);
     } else if (event is DeleteMeal) {
       yield* _deleteMeal(event);
     }
   }
 
   Stream<RecipieState> _deleteMeal(DeleteMeal event) async* {
-    final recipie = (state as RecipieLoadSuccess);
+    final recipie = (state as RecipieLoadSuccess).recipie;
     yield RecipieLoading();
     try {
       final mealId = event.mealId;
-      await recipieItemRepository.deleteItem(mealId);
-      recipie.recipieMeals.removeWhere((meal) => meal.id == mealId);
-      yield RecipieLoadSuccess(recipie.recipieMeals, recipie.recipieId);
+      if (recipie.meals.length < 0)
+        RecipieLoadFailure('CANNOT DELETE THE LAST MEAL ITEM');
+      else {
+        await recipieItemRepository.deleteItem(mealId);
+        recipie.recipieMeals.removeWhere((meal) => meal.id == mealId);
+      }
+      yield RecipieLoadSuccess(recipie);
     } catch (e) {
       RecipieLoadFailure('CANNOT DELELETE');
     }
@@ -60,41 +62,89 @@ class RecipieBloc extends Bloc<RecipieEvent, RecipieState> {
         final userMeals = (mealBloc.state as MealLoadSuccess).myMeals;
         final recipie =
             await recipieRepository.findItem(event.recipieId, userMeals);
-        // Recipie.recipieMealsToItemMeals(userMeals, recipie.meals);
-        //TODO: CONVERT RECPIEMELAS INTO ITEMMELAS
-        yield RecipieLoadSuccess(recipie.meals, event.recipieId);
+        yield RecipieLoadSuccess(recipie);
       } catch (e) {
         print(e);
       }
     }
   }
 
-  Stream<RecipieState> _addMeal(AddMeal event) async* {
-    final oldMeals = (state as RecipieLoadSuccess).recipieMeals;
+  Stream<RecipieState> _addEditMeal(AddMeal event) async* {
+    final recipie = (state as RecipieLoadSuccess).recipie;
     yield RecipieLoading();
     try {
-      //final meal = await mealItemRepository.addItem(event.meal);
-      //oldMeals.add(meal);
-      // yield RecipieLoadSuccess(oldMeals);
-    } catch (_) {
-      print(_);
-      RecipieLoadFailure('SOMETHING WENT WRONG');
+      final newMeal = event.meal;
+      final isNewRecipie = recipie.id == '';
+      Recipie newRecipie;
+      final recipieId = uuidd.v4();
+      if (isNewRecipie)
+        newRecipie = await recipieRepository.addItem(recipie, recipieId);
+      else
+        newRecipie = recipie;
+      final recipieIdMeal = uuidd.v4();
+
+      final trackMealItem = RecipieItem(
+          id: recipieIdMeal,
+          mealId: newMeal.id,
+          recipieId: newRecipie.id,
+          qty: newMeal.servingSize);
+
+      //Old Data
+      final recipieMeals = newRecipie.meals;
+      var oldCarbs = newRecipie.macrosConsumed.carbs;
+      var oldFats = newRecipie.macrosConsumed.fats;
+      var oldProtein = newRecipie.macrosConsumed.protein;
+
+      //New Meal data
+      // final newMeal = event.meal;
+      final mealProtein = newMeal.protein;
+      final mealCarbs = newMeal.carbs;
+      final mealFats = newMeal.fats;
+
+      // Meal Item
+      MealItem mealFoundInRecipie;
+
+      //Replacement
+      final indexMeal = recipieMeals.indexWhere((m) => m.id == newMeal.id);
+      if (indexMeal != -1) mealFoundInRecipie = recipieMeals[indexMeal];
+
+      Macro newMacros;
+      if (mealFoundInRecipie != null) {
+        // JUST UPDATE
+        final oldMealCarbs = mealFoundInRecipie.carbs;
+        final oldMealFats = mealFoundInRecipie.fats;
+        final oldMealProtein = mealFoundInRecipie.protein;
+        newMacros = Macro(
+            oldProtein - oldMealProtein + mealProtein,
+            oldCarbs - oldMealCarbs + mealCarbs,
+            oldFats - oldMealFats + mealFats);
+        await recipieItemRepository.updateItem(trackMealItem);
+        final indexNewMeal =
+            recipieMeals.indexWhere((meal) => meal.id == newMeal.id);
+        recipieMeals[indexNewMeal] = newMeal;
+      } else {
+        await recipieItemRepository.addItem(trackMealItem);
+        newMacros = Macro(
+            oldProtein + mealProtein, oldCarbs + mealCarbs, oldFats + mealFats);
+      }
+      await recipieRepository.updateItem(newRecipie, newMacros);
+
+      //  Yielding results
+      final recipieFinal = Recipie(
+          id: newRecipie.id,
+          recipeMeal: newRecipie.recipeMeal,
+          macrosConsumed: newMacros,
+          meals: recipieMeals);
+      yield RecipieLoadSuccess(recipieFinal);
+    } catch (e) {
+      print(e);
+      TrackLoadedFailure('WENT GRONG');
     }
   }
 
-  Stream<RecipieState> _editMeal(UpdateMeal event) async* {
-    final prevMeals = (state as RecipieLoadSuccess).recipieMeals;
-    yield RecipieLoading();
-    try {
-      final updatedMeal = event.meal;
-      // await mealItemRepository.updateItem(updatedMeal);
-      final indexUpdatedMeal =
-          prevMeals.indexWhere((m) => m.id == updatedMeal.id);
-      prevMeals[indexUpdatedMeal] = updatedMeal;
-
-      //  yield RecipieLoadSuccess(prevMeals);
-    } catch (e) {
-      RecipieLoadFailure('CANNOT UPDATE MEAL');
-    }
+  @override
+  Future<void> close() {
+    mealTrackGroupSubscription.cancel();
+    return super.close();
   }
 }
